comm

+++ database/migrations/2024_05_22_093634_create_orders_table.php
@@ -8,17 +8,19 @@ class CreateOrdersTable extends Migration
 {
     public function up()
     {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('phone');
-            $table->string('address');
-            $table->string('payment_method');
-            $table->date('buy_date');
-            $table->string('status');
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('orders')) {
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('phone');
+                $table->string('address');
+                $table->string('payment_method');
+                $table->date('buy_date');
+                $table->string('status');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
